refactor(agents): move live model policy to test owner
