+++ src/gateway/gateway-models.profiles.live.test.ts
@@ -32,19 +32,6 @@ import type { AuthProfileStore } from "../agents/auth-profiles/types.js";
 import { collectProviderApiKeys } from "../agents/live-auth-keys.js";
 import { isModelNotFoundErrorMessage } from "../agents/live-model-errors.js";
 import {
-  DEFAULT_HIGH_SIGNAL_LIVE_MODEL_LIMIT,
-  DEFAULT_SMALL_LIVE_MODEL_LIMIT,
-  getHighSignalLiveModelPriorityIndex,
-  isHighSignalLiveModelRef,
-  isSmallLiveModelRef,
-  listPrioritizedHighSignalLiveModelRefs,
-  listPrioritizedSmallLiveModelRefs,
-  resolveHighSignalLiveModelLimit,
-  selectHighSignalLiveItems,
-  selectSmallLiveItems,
-  shouldExcludeProviderFromDefaultHighSignalLiveSweep,
-} from "../agents/live-model-filter.js";
-import {
   isLiveProfileKeyModeEnabled,
   isLiveTestEnabled,
   readLiveTestConfig,
@@ -59,7 +46,20 @@ import { normalizeProviderId } from "../agents/model-selection.js";
 import { shouldSuppressBuiltInModelCore } from "../agents/model-suppression.js";
 import { ensureOpenClawModelsJson } from "../agents/models-config.js";
 import { STREAM_ERROR_FALLBACK_TEXT } from "../agents/stream-message-shared.js";
-import { appendPrioritizedDynamicLiveModels } from "../agents/test-helpers/live-model-dynamic-candidates.js";
+import {
+  appendPrioritizedDynamicLiveModels,
+  DEFAULT_HIGH_SIGNAL_LIVE_MODEL_LIMIT,
+  DEFAULT_SMALL_LIVE_MODEL_LIMIT,
+  getHighSignalLiveModelPriorityIndex,
+  isHighSignalLiveModelRef,
+  isSmallLiveModelRef,
+  listPrioritizedHighSignalLiveModelRefs,
+  listPrioritizedSmallLiveModelRefs,
+  resolveHighSignalLiveModelLimit,
+  selectHighSignalLiveItems,
+  selectSmallLiveItems,
+  shouldExcludeProviderFromDefaultHighSignalLiveSweep,
+} from "../agents/test-helpers/live-model-dynamic-candidates.js";
 import { createLiveTargetMatcher } from "../agents/test-helpers/live-target-matcher.js";
 import { mergeWorkspaceSetupState } from "../agents/workspace-state-store.js";
 import { ensureAgentWorkspace } from "../agents/workspace.js";
